added versioning for api added testcontainers rewrite app test using restassuredmockmvc
