+++ build.gradle
@@ -17,12 +17,14 @@ dependencies {
   implementation "org.springframework.boot:spring-boot-starter-web"
   implementation "org.springframework.boot:spring-boot-starter-data-jpa"
   implementation "mysql:mysql-connector-java:${mysqlConnectorVersion}"
+  implementation platform("org.testcontainers:testcontainers-bom:${testcontainerVersion}")
   compileOnly "org.projectlombok:lombok"
   annotationProcessor "org.projectlombok:lombok"
-  testImplementation "io.rest-assured:rest-assured:${restAssuredVersion}"
+  testImplementation "io.rest-assured:spring-mock-mvc:${restAssuredVersion}"
   testImplementation "io.rest-assured:json-schema-validator:${restAssuredVersion}"
   testImplementation "org.springframework.boot:spring-boot-starter-test"
-  testImplementation "com.h2database:h2"
+  testImplementation "org.testcontainers:junit-jupiter:${testcontainerVersion}"
+  testImplementation("org.testcontainers:mysql")
 }
 
 tasks.named('test') {
