+++ gradle.properties
@@ -1,2 +1,3 @@
 restAssuredVersion=5.3.0
 mysqlConnectorVersion=8.0.33
+testcontainerVersion=1.18.0
